Modify how objects are created from strings
I originally wanted constructors to accept a single string
representation of the object, but have abandoned this idea in favor of a
more explicit approach where class.fromString() is used.

As more constructor methods are added to the classes, this will provide
a uniform experience, where constructors are decomposed and efficient,
while the static from* methods create easier entry points.