Add implicit octave shift for pitch class comparison
This is a subtle, but important change. If we compare the difference
between G4 and C5, the answer is obvious. But if we compare the pitch
classes G and C, it can be interpreted both as G4 to C4 or G4 to C5.

I've changed the default interpretation from the former to the latter,
as it feels more intuitive that the first pitch class given is the one
with the lowest pitch.

This change also enforces that pitches and pitch classes can't be
compared.