Add support for notelists with both pitches and pitch classes
I want this to be supported, as it can be useful, but there are some
limitations. As we can't calculate intervals, it's impossible to search
for matching chords or scales, at least without guessing what the user
means.

For now, I throw an error telling the user to explicitly convert to a
list of pitches or pitch classes depending on what they are trying to
do. Maybe I can think of a solution to this later, but I suspect the
problem is inherently ambigious.