Modify ensureType to check for strings rather than objects
This was originally motivated by an error when working in a code base
with both a kamasi dependency and an indirect kamasi dependency.
instanceof no longer works, as the object was created in another
execution context.

This commit does not solve that problem per se, but I think it makes
more sense. It only checks for the type we know we can convert, then
fails explicitly if we try to create an object with another type.