Move Scales and Chords to NoteList class
This is a pretty big change, but I didn't feel that the Scale and Chord
classes did enough to justify standing by themselves. Inheritance adds
complexity, and these two were essentially just a helper function.

If more functionality is introduced, a reversal may be considered. For
now I think this is the way to go.