Modify search functionality
Another change in how this works, I think this is the one. A search is
now performed lazily by calling e.g. notes().chords().exact() or
notes().scales().supersets()

This way works well for Diatonic, but I'm still not sure if it's better
the other way around, or by using attributes. I like how this follows
the chaining logic of the rest of the API, even though the steps are
kind of useless.