Replace `eslint` with `biome` (lint only)
`biome` is newer, faster, shinier, and notably supports TypeScript
out of the box (which we will soon need).

I opted against using the formatter in this project, however. Music
theory is full of patterns I want to reflect in the code even when
it breaks formatting standards.

Maybe I'll introduce it later and explicitly ignore some files or
lines, but for now I just keep it disabled.